factor out getFreePort
